TEIID-3899 moving to a full import property
