+++ connectors/translator-ws/src/main/resources/org/teiid/translator/ws/i18n.properties
@@ -28,4 +28,4 @@ TEIID15003=Unsupported binding found. Only SOAP11, SOAP12, HTTP bindings are sup
 TEIID15004=Operation {0} is RPC style; only Document style is supported. Procedure skipped.
 TEIID15005=Error Calling HTTP Service - {0} {1}
 TEIID15006=Invalid JSON for header.  The value must be a JSON Object containing list of primitive or primitive values.
-TEIID15007=Connection to web service failed; Connection is required to import WSDL metadata.
+TEIID15007=Connection to web service failed; Connection is required to import WSDL metadata.  Use importer.importWSDL=false if the WSDL is not needed.
